小数が登録できるように変更

+++ db/migrate/20230710122525_create_foods.rb
@@ -2,9 +2,9 @@ class CreateFoods < ActiveRecord::Migration[7.0]
   def change
     create_table :foods do |t|
       t.string :name, null: false
-      t.integer :protein, null: false
-      t.integer :fat, null: false
-      t.integer :carbohydrate, null: false
+      t.float :protein, null: false
+      t.float :fat, null: false
+      t.float :carbohydrate, null: false
       t.references :user, null: false, foreign_key: true
       t.timestamps
     end
